+++ configure.ac
@@ -44,6 +44,7 @@ GLIB_MIN_VERSION="2.26.1"
 OPENSSL_MIN_VERSION="0.9.8"
 LIBDBI_MIN_VERSION="0.9.0"
 LIBRABBITMQ_MIN_VERSION="0.5.3"
+LIBRDKAFKA_MIN_VERSION="0.11.3"
 IVYKIS_MIN_VERSION="0.36.1"
 IVYKIS_UPDATED_VERSION="0.39"
 JSON_C_MIN_VERSION="0.9"
@@ -256,6 +257,10 @@ AC_ARG_ENABLE(json,
 AC_ARG_ENABLE(amqp,
           [  --enable-amqp        Enable amqp destination (default: auto)]
               ,,enable_amqp="auto")
+
+AC_ARG_ENABLE(kafka,
+          [  --enable-kafka        Enable kafka destination (default: auto)]
+              ,,enable_kafka="auto")
 
 AC_ARG_ENABLE(stomp,
           [  --enable-stomp        Enable stomp destination (default: yes)]
@@ -1307,6 +1312,16 @@ if test "$enable_amqp" = "yes" && test "$with_librabbitmq_client" = "no"; then
        AC_MSG_ERROR([Could not find librabbitmq-c, and AMQP support was explicitly enabled.])
 fi
 
+dnl ***************************************************************************
+dnl rdkafka headers/libraries
+dnl ***************************************************************************
+
+PKG_CHECK_MODULES(LIBRDKAFKA, rdkafka >= $LIBRDKAFKA_MIN_VERSION, with_librdkafka="yes", with_librdkafka="no")
+
+if test "$enable_kafka" = "yes" && test "$with_librdkafka" = "no"; then
+       AC_MSG_ERROR([Could not find librdkafka, and Kafka support was explicitly enabled.])
+fi
+
 
 if test "x$enable_native" = "xyes" -o "x$enable_native" = "xauto"; then
     AC_CONFIG_FILES([syslog-ng-native-connector.pc])
@@ -1560,6 +1575,15 @@ if test "x$enable_amqp" = "xauto"; then
     AC_MSG_RESULT([$enable_amqp])
 fi
 
+if test "x$enable_kafka" = "xauto"; then
+    AC_MSG_CHECKING(whether to enable kafka source support)
+    if test "x$with_librdkafka" != "xno"; then
+        enable_kafka="yes"
+    else
+        enable_kafka="no"
+    fi
+    AC_MSG_RESULT([$enable_kafka])
+fi
 
 if test "x$enable_systemd" = "xauto"; then
 	if test "$ostype" = "Linux" -a "$have_libsystemd" = "yes"; then
@@ -1799,6 +1823,7 @@ AM_CONDITIONAL(ENABLE_MONGODB, [test "$enable_mongodb" = "yes"])
 AM_CONDITIONAL(ENABLE_SMTP, [test "$enable_smtp" = "yes"])
 AM_CONDITIONAL(ENABLE_HTTP, [test "$enable_http" = "yes"])
 AM_CONDITIONAL(ENABLE_AMQP, [test "$enable_amqp" = "yes"])
+AM_CONDITIONAL(ENABLE_KAFKA, [test "$enable_kafka" = "yes"])
 AM_CONDITIONAL(ENABLE_STOMP, [test "$enable_stomp" = "yes"])
 AM_CONDITIONAL(ENABLE_JSON, [test "$enable_json" = "yes"])
 AM_CONDITIONAL(ENABLE_GEOIP, [test "$enable_geoip" = "yes"])
@@ -1938,6 +1963,7 @@ echo "  JSON support (module)       : ${enable_json:=no}"
 echo "  SMTP support (module)       : ${enable_smtp:=no}"
 echo "  HTTP support (module)       : ${enable_http:=no}"
 echo "  AMQP destination (module)   : ${enable_amqp:=no}"
+echo "  Kafka source (module)       : ${enable_kafka:=no}"
 echo "  STOMP destination (module)  : ${enable_stomp:=no}"
 echo "  GEOIP support (module)      : ${enable_geoip:=no}"
 echo "  GEOIP2 support (module)     : ${enable_geoip2:=no}"
