+++ modules/CMakeLists.txt
@@ -27,6 +27,7 @@ add_subdirectory(http)
 add_subdirectory(java)
 add_subdirectory(java-modules)
 add_subdirectory(json)
+add_subdirectory(kafka)
 add_subdirectory(kvformat)
 add_subdirectory(linux-kmsg-format)
 add_subdirectory(map-value-pairs)
